+++ Jenkinsfile
@@ -14,14 +14,14 @@ pipeline {
 
             git branch: 'main', 
                credentialsId: env.GIT_CREDENTIAL_ID,
-               url: 'https://github.com/Uniandes-isis2603/' + env.GIT_REPO
+               url: 'https://github.com/UDFJDC-ModelosProgramacion/' + env.GIT_REPO
          }
       }
       stage('GitInspector') { 
          steps {
             withCredentials([usernamePassword(credentialsId: env.GIT_CREDENTIAL_ID, passwordVariable: 'GIT_PASSWORD', usernameVariable: 'GIT_USERNAME')]) {
                sh 'mkdir -p code-analyzer-report'
-               sh """ curl --request POST --url https://code-analyzer.virtual.uniandes.edu.co/analyze --header "Content-Type: application/json" --data '{"repo_url":"git@github.com:Uniandes-isis2603/${GIT_REPO}.git", "access_token": "${GIT_PASSWORD}" }' > code-analyzer-report/index.html """   
+               sh """ curl --request POST --url https://code-analyzer.virtual.udistrital.edu.co/analyze --header "Content-Type: application/json" --data '{"repo_url":"git@github.com:UDFJDC-ModelosProgramacion/${GIT_REPO}.git", "access_token": "${GIT_PASSWORD}" }' > code-analyzer-report/index.html """   
             }
             publishHTML (target: [
                allowMissing: false,
@@ -78,7 +78,7 @@ pipeline {
                   sh '''
                      java -version
                      rsync --recursive . bookstore-back
-                     java -cp /eclipse/plugins/org.eclipse.equinox.launcher_1.5.700.v20200207-2156.jar org.eclipse.equinox.launcher.Main -application co.edu.uniandes.archtoring.archtoring bookstore-back
+                     java -cp /eclipse/plugins/org.eclipse.equinox.launcher_1.5.700.v20200207-2156.jar org.eclipse.equinox.launcher.Main -application co.edu.udistrital.archtoring.archtoring bookstore-back
                   '''
                }
             }
