+++ src/main/java/co/edu/udistrital/mdp/ZZZ/exceptions/EntityNotFoundException.java
@@ -1,0 +1,13 @@
+package co.edu.udistrital.mdp.ZZZ.exceptions;
+
+/*
+ * Excepción que se lanza cuando en el proceso de búsqueda no se encuenta una entidad
+ */
+public class EntityNotFoundException extends Exception {
+	
+	private static final long serialVersionUID = 1L;
+	
+	public EntityNotFoundException(String message) {
+		super(message);
+	}
+}
