+++ src/main/java/co/edu/udistrital/mdp/ZZZ/exceptions/IllegalOperationException.java
@@ -1,0 +1,14 @@
+package co.edu.udistrital.mdp.ZZZ.exceptions;
+
+/*
+ * Excepción que se lanza cuando se realiza una operación ilegal
+ */
+public class IllegalOperationException extends Exception {
+
+	private static final long serialVersionUID = 1L;
+
+	public IllegalOperationException(String message) {
+		super(message);
+	}
+}
+
