+++ src/main/java/co/edu/udistrital/mdp/ZZZ/exceptions/RestExceptionHandler.java
@@ -1,4 +1,4 @@
-package co.edu.uniandes.dse.ZZZ.exceptions;
+package co.edu.udistrital.mdp.ZZZ.exceptions;
 
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.HttpStatus.PRECONDITION_FAILED;
