--- src/main/java/co/edu/uniandes/dse/ZZZ/exceptions/EntityNotFoundException.java
@@ -1,36 +1,0 @@
-/*
-MIT License
-
-Copyright (c) 2021 Universidad de los Andes - ISIS2603
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-*/
-package co.edu.uniandes.dse.ZZZ.exceptions;
-
-/*
- * Excepción que se lanza cuando en el proceso de búsqueda no se encuenta una entidad
- */
-public class EntityNotFoundException extends Exception {
-	
-	private static final long serialVersionUID = 1L;
-	
-	public EntityNotFoundException(String message) {
-		super(message);
-	}
-}
